refactor: Updated folder structure (#18)
