+++ frontend-service/src/app/css/index.css
@@ -1,6 +1,6 @@
 body {
   margin: 0;
-  font-family: 
+  font-family:
     -apple-system,
     BlinkMacSystemFont,
     'Segoe UI',
