+++ frontend-service/src/app/js/App.js
@@ -1,6 +1,6 @@
 import React from "react";
-import logo from "./logo.svg";
-import "./App.css";
+import logo from "../../assets/logo.svg";
+import "../css/App.css";
 
 function App() {
   return (
@@ -8,7 +8,7 @@ function App() {
       <header className="App-header">
         <img src={logo} className="App-logo" alt="logo" />
         <p>
-          Edit <code>src/App.js</code> and save to reload.
+          Edit <code>src/js/App.js</code> and save to reload.
         </p>
         <a
           className="App-link"
