+++ frontend-service/src/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import "./index.css";
-import App from "./App";
+import "./app/css/index.css";
+import App from "./app/js/App";
 import * as serviceWorker from "./serviceWorker";
 
 ReactDOM.render(
